moved the overwrite of the NIST enthalpy equation to natural_gas_props.py, renamed and reorganized files
